finished hw4-8

+++ ELibrary/posts/views.py
@@ -3,6 +3,7 @@ from django.http.response import HttpResponse, JsonResponse
 from django.views.decorators.http import require_GET, require_POST
 from datetime import datetime
 from django.shortcuts import get_object_or_404
+from application.decorators import login_required
 
 from users.models import User
 from posts.models import Post, Category 
@@ -27,20 +28,29 @@ def create_post(request):
     return JsonResponse({'created': dct})
 
 
+@login_required
 @require_GET
 def delete_post(request, pk):
     post = get_object_or_404(Post, pk=pk)
+    
+    if post.author.username != request.user.username:
+        return HttpResponse(status=404)
+
     post.delete()
     return JsonResponse({'deleted': {'post_id': pk}})
 
 
+@login_required
 @require_GET
 def get_post_info(request, pk):
     post = get_object_or_404(Post, pk=pk)
 
+    if post.author.username != request.user.username:
+        return HttpResponse(status=404)
+
     info = {'pk': pk,
             'title': post.title,
-            'author': post.author.email,
+            'author': post.author.username,
             'pub_datetime': post.pub_datetime,
             'body': post.body,}
 
@@ -64,12 +74,15 @@ def update_post(request):
     return JsonResponse({'updated': info})
 
 
+@login_required
 @require_GET
 def get_list_of_posts(request):
-    posts = Post.objects.all()
+    posts = Post.objects.filter(author = request.user)
+
 
     info = {post.pk: {'title': post.title,
-                      'author': post.author.email,
+                      'author': post.author.username,
+                      'author_email': post.author.email,
                       'pub_datetime': post.pub_datetime,
                       'body': post.body,} for post in posts}
